requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==0.4.6
altair==4.2.0
altgraph==0.17.3
antlr4-python3-runtime==4.8
anyio==3.6.2
async-timeout==4.0.2
attrs==22.1.0
bcrypt==4.0.1
blinker==1.5
brotlipy @ file:///Users/runner/miniforge3/conda-bld/brotlipy_1648854242877/work
cachetools==5.2.0
certifi==2022.6.15
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1656782892388/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1655906222726/work
click==8.1.3
-e git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1#egg=clip
coloredlogs==15.0.1
commonmark==0.9.1
ConfigArgParse==1.5.3
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography_1652967108255/work
cycler==0.11.0
decorator==5.1.1
diffusers==0.2.4
einops==0.3.0
entrypoints==0.4
fastapi==0.88.0
ffmpy==0.3.0
filelock==3.8.0
flask==2.0.1
flatbuffers==2.0.7
fonttools==4.33.3
frozenlist==1.3.1
fsspec==2022.8.2
ftfy==6.1.1
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.9.1
google-auth-oauthlib==0.4.6
gradio==3.13.0
grpcio==1.48.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.1
huggingface-hub==0.9.1
humanfriendly==10.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1642433548627/work
imageio==2.9.0
imageio-ffmpeg==0.4.2
imgaug==0.4.0
importlib-metadata==4.12.0
invisible-watermark==0.1.5
Jinja2==3.1.2
jsonschema==4.15.0
kiwisolver==1.4.3
kornia==0.6.0
linkify-it-py==1.0.3
lpips==0.1.4
macholib==1.16.2
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.5.2
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.2.1
multidict==6.0.2
networkx==2.8.5
numpy==1.23.1
oauthlib==3.2.0
omegaconf==2.1.1
onnx==1.12.0
onnxruntime==1.12.1
opencv-python==4.6.0.66
orjson==3.8.3
packaging==21.3
pandas==1.5.0rc0
paramiko==2.12.0
Pillow @ file:///Users/runner/miniforge3/conda-bld/pillow_1657007219741/work
plyfile==0.7.4
protobuf==3.19.4
pudb==2019.2
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pycryptodome==3.16.0
pydantic==1.10.2
pydeck==0.8.0b1
pyDeprecate==0.3.1
pydub==0.25.1
Pygments==2.13.0
pyinstaller @ file:///Users/zain/workspace/pyinstaller
pyinstaller-hooks-contrib==2022.14
Pympler==1.0.1
PyNaCl==1.5.0
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1643496850550/work
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks @ file:///Users/runner/miniforge3/conda-bld/pysocks_1648857374584/work
python-dateutil==2.8.2
python-multipart==0.0.5
pytorch-lightning==1.4.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==6.0
regex==2022.8.17
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1656534056640/work
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==12.5.1
rsa==4.9
scikit-image==0.19.3
scipy==1.9.0
semver==3.0.0.dev3
shapely==2.0a1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.22.0
streamlit==1.12.2
sympy==1.11.1
-e git+https://github.com/CompVis/taming-transformers.git@24268930bf1dce879235a7fddd0b2355b84d7ea6#egg=taming_transformers
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
test-tube==0.7.5
tifffile==2022.7.28
tokenizers==0.12.1
toml==0.10.2
toolz==0.12.0
torch==1.13.0.dev20220705
torch-fidelity==0.3.0
torchmetrics==0.6.0
torchvision==0.14.0.dev20220705
tornado==6.2
tqdm==4.64.0
transformers==4.19.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1656706066251/work
tzdata==2022.2
tzlocal==4.2
uc-micro-py==1.0.1
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1647489083693/work
urwid==2.1.2
uvicorn==0.20.0
validators==0.20.0
wcwidth==0.2.5
websockets==10.4
Werkzeug==2.2.1
yarl==1.8.1
zipp==3.8.1
